[frontend] lint debug
