+++ frontend/src/components/Annotation/SelectActiveLearning.tsx
@@ -5,8 +5,8 @@ import { PiEmptyBold } from 'react-icons/pi';
 import Select from 'react-select';
 import { Tooltip } from 'react-tooltip';
 import { useRetrainQuickModel, useTrainQuickModel } from '../../core/api';
-import { useAppContext } from '../../core/context';
 import { useNotifications } from '../../core/notifications';
+import { useAppContext } from '../../core/useAppContext';
 import { getRandomName, sortDatesAsStrings } from '../../core/utils';
 import { ActiveModel } from '../../types';
 import { ButtonNewFeature } from '../ButtonNewFeature';
