+++ frontend/src/components/FeaturesManagement.tsx
@@ -3,7 +3,7 @@ import { FC, useState } from 'react';
 import { Modal } from 'react-bootstrap';
 import { useParams } from 'react-router-dom';
 import { useDeleteFeature, useGetFeatureInfo, useResetFeatures } from '../core/api';
-import { useAppContext } from '../core/context';
+import { useAppContext } from '../core/useAppContext';
 import { sortDatesAsStrings } from '../core/utils';
 import { FeatureDescriptionModelOut } from '../types';
 import { ButtonNewFeature } from './ButtonNewFeature';
