+++ frontend/src/components/SchemesManagement.tsx
@@ -7,8 +7,8 @@ import { IoDuplicate } from 'react-icons/io5';
 import { MdDriveFileRenameOutline } from 'react-icons/md';
 import Select from 'react-select';
 import { useAddScheme, useDeleteScheme, useDuplicateScheme, useRenameScheme } from '../core/api';
-import { useAppContext } from '../core/context';
 import { useNotifications } from '../core/notifications';
+import { useAppContext } from '../core/useAppContext';
 import { getRandomName } from '../core/utils';
 import { SchemeModel } from '../types';
 /*
