+++ frontend/src/components/layout/ProjectActionsSideBar.tsx
@@ -8,8 +8,8 @@ import { PiTagDuotone } from 'react-icons/pi';
 import { RiAiGenerate } from 'react-icons/ri';
 import { Link } from 'react-router-dom';
 import { useGetServer } from '../../core/api';
-import { useAuth } from '../../core/auth';
 import { useNotifications } from '../../core/notifications';
+import { useAuth } from '../../core/useAuth';
 import { ProjectStateModel } from '../../types';
 import { AnnotationIcon } from '../Icons';
 import { PossibleProjectActions } from './ProjectPageLayout';
