+++ frontend/src/pages/AccountPage.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { ChangePassword } from '../components/forms/ChangePassword';
 import { PageLayout } from '../components/layout/PageLayout';
-import { useAuth } from '../core/auth';
+import { useAuth } from '../core/useAuth';
 
 export const AccountPage: FC = () => {
   const { authenticatedUser } = useAuth();
