+++ frontend/src/pages/ProjectTagPage.tsx
@@ -2,7 +2,7 @@ import { FC, useEffect, useState } from 'react';
 import Tab from 'react-bootstrap/Tab';
 import Tabs from 'react-bootstrap/Tabs';
 import { useParams } from 'react-router-dom';
-import { useAppContext } from '../core/context';
+import { useAppContext } from '../core/useAppContext';
 
 import { useLocation } from 'react-router-dom';
 import { AnnotationDisagreementManagement } from '../components/Annotation/AnnotationDisagreementManagement';
